# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "handlers.neighborhoodEdge": {
        "properties": {
          "edge_type": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "target": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.neighborhoodNode": {
        "properties": {
          "collection": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "title": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.neighborhoodRequest": {
        "properties": {
          "depth": {
            "type": "integer"
          },
          "direction": {
            "type": "string"
          },
          "edge_types": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "focus": {
            "type": "string"
          },
          "node_kind": {
            "type": "string"
          },
          "workspace": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "handlers.neighborhoodResponse": {
        "properties": {
          "edges": {
            "items": {
              "$ref": "#/components/schemas/handlers.neighborhoodEdge"
            },
            "type": "array"
          },
          "frontier_nodes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "node_kind": {
            "type": "string"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/handlers.neighborhoodNode"
            },
            "type": "array"
          },
          "truncated": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
      "handlers.overviewRequest": {
        "properties": {
          "edge_types": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "limit": {
            "type": "integer"
          },
          "mode": {
            "type": "string"
          },
          "workspace": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "CSRFToken": {
        "description": "Required on write endpoints; enforced by middleware.CSRF.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "WorkspaceAuth": {
        "description": "Workspace hash via query param or JSON body; resolved by workspaceMiddleware.",
        "in": "query",
        "name": "workspace",
        "type": "apiKey"
      },
      "WorkspaceRegisteredAuth": {
        "description": "Workspace must be already-registered (write-path gate); enforced by workspaceRegisteredMiddleware.",
        "in": "query",
        "name": "workspace",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Self-describing OpenAPI 3.0 spec for nano-brain's non-MCP REST API",
    "title": "nano-brain REST API",
    "version": "1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/graph/neighborhood": {
      "post": {
        "description": "Breadth-first traverses the graph around a focus node up to a given depth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.neighborhoodRequest"
              }
            }
          },
          "description": "Neighborhood query",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.neighborhoodResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "security": [
          {
            "WorkspaceRegisteredAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "BFS graph neighborhood around a focus node",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/overview": {
      "post": {
        "description": "Returns the top-N most-connected nodes for the workspace and all edges between them",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/handlers.overviewRequest"
              }
            }
          },
          "description": "Overview query",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/handlers.neighborhoodResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceAuth": []
          }
        ],
        "summary": "Top-N most-connected graph nodes overview",
        "tags": [
          "graph"
        ]
      }
    },
    "/api/v1/graph/pagerank/compute": {
      "post": {
        "description": "Recomputes and persists PageRank centrality scores over the workspace's call-edge graph",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "WorkspaceRegisteredAuth": []
          },
          {
            "CSRFToken": []
          }
        ],
        "summary": "Recompute PageRank scores for the call graph",
        "tags": [
          "graph"
        ]
      }
    },
...
  }
}